+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 * Change README, as v0.2.0 is already out and under testing.
+* magefile niceties: "build" target for native OS, ".exe" extension for windows target, more precise version info from git
 
 ### Fixed